feat(cache): 新增缓存删除与hash更新功能
新增 CommonCacheService 接口方法 delKey 和 hashFlushValue，
并在 CommonCacheServiceImpl 中实现。同时增加对应常量
DEL_KEY_ERROR 用于缓存删除失败提示。

fix(product): 优化商品详情缓存处理逻辑

完善商品详情获取流程中对缓存未命中情况的异常处理，
增加日志记录并抛出业务异常；调整异步回填缓存策略，
确保数据一致性。refactor(location): 修改 location 字段名为 locationCode

将 Location 类中的 location 属性重命名为 locationCode，
以更准确表达其含义，并同步修改相关 getter、构造函数及equals/hashCode 方法。feat(product): 增加获取商品信息用于缓存的方法

在 ProductMapper 和 ProductRepository 中添加
getProductForCache 方法，支持通过 productId 查询完整的
商品信息用于缓存重建。

refactor(mapper): 调整 gerConditionNameById 参数类型

将 gerConditionNameById 方法参数由 Long 改为 Integer，保持与数据库字段类型一致。

chore(utils):修正 RedisUtils 注释描述

更新 hashSet 方法上的注释内容，使其更加清晰地说明该方法
的作用是缓存 hash 类型的数据。

+++ product-module/src/main/java/com/zhuxi/product/module/application/service/process/PublishShProcess.java
@@ -33,7 +33,7 @@ public class PublishShProcess {
      * @return 业务需要的冗余字段
      */
     @Transactional(readOnly = true, propagation = Propagation.SUPPORTS)
-    public List<String> getOtherNames(Long conditionId, Long categoryId) {
+    public List<String> getOtherNames(Integer conditionId, Long categoryId) {
         String categoryName = repository.getCategoryNameById(categoryId);
         String conditionName = repository.gerConditionNameById(conditionId);
